feat: Rozbiecie subscribera
