+++ src/EventSubscriber/AuthValidationSubscriber.php
@@ -7,7 +7,6 @@ use App\Entity\User;
 use App\Exception\AuthenticationException;
 use App\Exception\DataNotFoundException;
 use App\Exception\PermissionException;
-use App\Exception\ResponseExceptionInterface;
 use App\Exception\TechnicalBreakException;
 use App\Repository\AuthenticationTokenRepository;
 use App\Repository\TechnicalBreakRepository;
@@ -17,9 +16,6 @@ use Doctrine\ORM\NonUniqueResultException;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
-use Symfony\Component\HttpKernel\Event\ExceptionEvent;
-use Symfony\Component\HttpKernel\Event\ResponseEvent;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\KernelEvents;
 
 /**
@@ -31,21 +27,18 @@ class AuthValidationSubscriber implements EventSubscriberInterface
     private AuthenticationTokenRepository $authenticationTokenRepository;
     private TechnicalBreakRepository $technicalBreakRepository;
     private UserRepository $userRepository;
-    private LoggerInterface $responseLogger;
     private LoggerInterface $requestLogger;
 
     public function __construct(
         AuthenticationTokenRepository $authenticationTokenRepository,
         TechnicalBreakRepository      $technicalBreakRepository,
         UserRepository                $userRepository,
-        LoggerInterface               $responseLogger,
         LoggerInterface               $requestLogger,
     )
     {
         $this->authenticationTokenRepository = $authenticationTokenRepository;
         $this->technicalBreakRepository = $technicalBreakRepository;
         $this->userRepository = $userRepository;
-        $this->responseLogger = $responseLogger;
         $this->requestLogger = $requestLogger;
     }
 
@@ -147,86 +140,10 @@ class AuthValidationSubscriber implements EventSubscriberInterface
         }
     }
 
-    /**
-     * @param ExceptionEvent $event
-     * @return void
-     */
-    public function onKernelException(ExceptionEvent $event): void
-    {
-        $exception = $event->getThrowable();
-
-        if ($exception instanceof ResponseExceptionInterface) {
-            $loggingContext = [
-                "statusCode" => $exception->getResponse()->getStatusCode(),
-                "file" => "[" . $exception->getLine() . "](" . $exception->getFile() . ")",
-                "responseData" => json_decode($exception->getResponse()->getContent(), true)
-            ];
-
-            $this->responseLogger->info("ResponseException", $loggingContext);
-
-            $event->setResponse($exception->getResponse());
-        } else {
-            $this->responseLogger->critical("ResponseException", ["class" => $exception::class, "data" => $exception]);
-
-            switch ($exception::class) {
-                case NotFoundHttpException::class:
-                {
-                    $notFoundException = new DataNotFoundException([$exception->getMessage()]);
-
-                    $event->setResponse($notFoundException->getResponse());
-                    break;
-                }
-            }
-        }
-    }
-
-    /**
-     * @throws NonUniqueResultException
-     */
-    public function onKernelResponse(ResponseEvent $event): void
-    {
-        $request = $event->getRequest();
-        $response = $event->getResponse();
-
-        $authorizationHeaderField = $request->headers->get("authorization");
-
-        $authToken = null;
-        if ($authorizationHeaderField !== null) {
-            $authToken = $this->authenticationTokenRepository->findActiveToken($authorizationHeaderField);
-        }
-
-        $technicalBreak = $this->technicalBreakRepository->findOneBy([
-            "active" => true
-        ]);
-
-        if($technicalBreak !== null){
-            $response->headers->set('Technical-Break', true);
-        }
-
-        $headersIterator = $response->headers->getIterator();
-
-        $loggerData = [
-            "requestUrl" => $request->getUri(),
-            "requestMethod" => $request->getMethod(),
-            "user" => $authToken?->getUser()->getId(),
-            "statusCode" => $response->getStatusCode(),
-            "headers" => $headersIterator->getArrayCopy(),
-            "responseData" => $response->getStatusCode() > 299 ? json_decode($response->getContent(), true) : null,
-        ];
-
-        if ($response->getStatusCode() > 499) {
-            $this->responseLogger->error("Response data", $loggerData);
-        } else {
-            $this->responseLogger->info("Response data", $loggerData);
-        }
-    }
-
     public static function getSubscribedEvents(): array
     {
         return [
             KernelEvents::CONTROLLER => 'onControllerCall',
-            KernelEvents::EXCEPTION => 'onKernelException',
-            KernelEvents::RESPONSE => 'onKernelResponse'
         ];
     }
